rename files to project-files
